update: Streamlined the process
